Add CoreWingCrossSectionMovement test module
Move WingCrossSectionMovement tests to a Core-level test module so they
directly test CoreWingCrossSectionMovement behavior. Keep a slim
WingCrossSectionMovement test module that verifies the public subclass
relationship and inherited behavior. Duplicate the five
WingCrossSectionMovement fixtures that consumer tests need since
WingMovement enforces the stricter public type.
