+++ tests/unit/__init__.py
@@ -20,6 +20,9 @@ This package contains the following modules:
 
     test_core.py: This module contains classes to test the core classes in _core.py.
 
+    test_core_wing_cross_section_movement.py: This module contains classes to test
+    CoreWingCrossSectionMovements.
+
     test_horseshoe_vortex.py: This module contains classes to test HorseshoeVortices.
 
     test_line_vortex.py: This module contains classes to test LineVortices.
@@ -71,6 +74,7 @@ import tests.unit.test_airfoil
 import tests.unit.test_airplane
 import tests.unit.test_airplane_movement
 import tests.unit.test_core
+import tests.unit.test_core_wing_cross_section_movement
 import tests.unit.test_horseshoe_vortex
 import tests.unit.test_line_vortex
 import tests.unit.test_movement
